add error handling to user actions

+++ src/models/findUser.js
@@ -3,7 +3,7 @@
 
 //import and create pool
 const pool = require('./util/pool');
-
+const { messageNoResults, checkNoResults } = require('./util/checkFind');
 
 const { formatUserOutput, formatContactOutput } = require('./util/formatOutput')
 
@@ -13,8 +13,10 @@ const findAllUsers = async () => {
     contact.phone, contact.address, contact.city, contact.state, contact.zip, contact.email \
     FROM users JOIN contact ON users.contact_id = contact.id ORDER BY id ASC'
   const results = await pool.query(sql);
+  const noResults = checkNoResults(results);
+  if (noResults) return noResults;
+
   const tableOutput = results.rows;
-  
   let contactObj;
   let user;
   const users = [];
@@ -29,41 +31,41 @@ const findAllUsers = async () => {
 
 //queries user using user ID and returns a user object with contact info OR false if user not in database.
 const findUserById = async (id) => {
-  const userRes = await pool.query('SELECT fullname, username, contact_id FROM users WHERE id = $1', [id])  //NOTE does not include password column
-  if (userRes.rows.length > 0) {
-  const userObj = userRes.rows[0]
-  const contactRes = await pool.query('SELECT * FROM contact WHERE id = $1', [userObj.contact_id])
-  const contactObj = contactRes.rows[0]
+  const userRes = await pool.query('SELECT fullname, username, contact_id FROM users WHERE id = $1', [id]);  //NOTE does not include password column
+  const noUserResults = checkNoResults(userRes);
+  if (noUserResults) return noUserResults;
+  const userObj = userRes.rows[0];
+  const contactRes = await pool.query('SELECT * FROM contact WHERE id = $1', [userObj.contact_id]);
+  const noContactResults = checkNoResults(contactRes);
+  if (noContactResults) return noContactResults;
+  const contactObj = contactRes.rows[0];
   const user = formatUserOutput(userObj, contactObj);
-
   return user;
-} else {
-  return false;
-}
 }
 
 //queries users to see if username is already in database an returns true if it is unique (not in database) OR false if the username already exists
 const isUsernameUnique = async (username) => {
-  const results = await pool.query('SELECT username FROM users WHERE username = $1', [username]) 
-  return results.rows.length === 0;
+  const results = await pool.query('SELECT username FROM users WHERE username = $1', [username]);
+  const noResults = checkNoResults(results);
+  return noResults;
 }
 
 //creates new user in database
 const addUser = async (newUser) => {
   const { fullname, username, password, contact } = newUser;
-  
   const { phone, address, city, state, zip, email } = contact;
   const contactSql = 'INSERT INTO contact (phone, address, city, state, zip, email) VALUES ($1, $2, $3, $4, $5, $6) RETURNING *';
   const contactRes = await pool.query(contactSql, [phone, address, city, state, zip, email]);
-
+  const noContactResults = checkNoResults(contactRes);
+  if (noContactResults) return noContactResults;
   const contactId = contactRes.rows[0].id;
   const userSql = 'INSERT INTO users (fullname, username, password, contact_id) VALUES ($1, $2, $3, $4) RETURNING *'; 
   const userRes = await pool.query(userSql, [fullname, username, password, contactId]);
-
+  const noUserResults = checkNoResults(userRes);
+  if (noUserResults) return noUserResults;
   const userObj = userRes.rows[0];
   const contactObj = contactRes.rows[0];
   const user = formatUserOutput(userObj, contactObj);
-
   return user;
 };
 
@@ -91,12 +93,14 @@ const changeUser = async (id, updates) => {
     const { fullname, username } = existingUser;
     const userSql = 'UPDATE users SET fullname = $1, username = $2 WHERE id = $3 RETURNING *' 
     const userRes = await pool.query(userSql, [fullname, username, id]);
-
+    const noUserResults = checkNoResults(userRes);
+    if (noUserResults) return noUserResults;
     const contactId = userRes.rows[0].contact_id;
     const { phone, address, city, state, zip, email } = existingUser.contact;
     const contactSql = 'UPDATE contact SET phone = $1, address = $2, city = $3, state = $4, zip = $5, email = $6 WHERE id = $7 RETURNING *';
     const contactRes = await pool.query(contactSql, [phone, address, city, state, zip, email, contactId]);
-
+    const noContactResults = checkNoResults(contactRes);
+    if (noContactResults) return noContactResults;
     const userObj = userRes.rows[0];
     const contactObj = contactRes.rows[0];
     const updatedUser = formatUserOutput(userObj, contactObj);
@@ -110,11 +114,14 @@ const removeUser = async (id) => {
   if (!existingUser) {
     return `No such user with id: ${id}`
   } else {
-    const deletedUser = await pool.query('DELETE FROM users WHERE id = $1 RETURNING *', [id]);
-    const contactId = deletedUser.contactId;
-    const deletedContact = await pool.query('DELETE FROM contact WHERE id = $1', [contactId]); //IDEA add a mark for deletion boolean column to contact table, set it to true instead of deleting.
-
-    const deletedUserObj = formatUserOutput(deletedUser, deletedContact);
+    const userRes = await pool.query('DELETE FROM users WHERE id = $1 RETURNING *', [id]);
+    const noUserResults = checkNoResults(userRes);
+    if (noUserResults) return noUserResults;
+    const contactId = userRes.contact_id;
+    const contactRes = await pool.query('DELETE FROM contact WHERE id = $1', [contactId]); //IDEA add a mark for deletion boolean column to contact table, set it to true instead of deleting.
+    const noContactResults = checkNoResults(contactRes);
+    if (noContactResults) return noContactResults;
+    const deletedUserObj = formatUserOutput(userRes.rows[0], contactRes.rows[0]);
     return deletedUserObj;
   }
 }
